Prioritize alignSelf over absolute positioning
The initial implementation of absolute positioning would follow a more free
approach of "if you choose absolute positioning, alignment should be done
through the left, top, right and bottom properties".

But what happens if the item also specifies `alignSelf==flex_end`, for instance?
With this change we prioritize the `alignSelf` value over the corresponding
cross-axis values (left/right, top/bottom, depending on the flexDirection).

Margins can still be used in this case to seaparte elements from the sides of
the container, if needed.